Implement greatly expanded support for g-code.
Moved gcode_common.h to gcode.cpp and started implementing support for more gcode.

Committing the deletion of gcode_common.h.

Initial implementation of g-code circular arcs.

This implements support for the G2 command, using the ij style of
arguments. Circular arcs are approximated using cubic bezier curves
as supported by the cricut, using an implementation from

APPROXIMATION OF A CUBIC BEZIER CURVE BY CIRCULAR ARCS AND VICE VERSA,
Aleksas Riškus, ISSN 1392 – 124X INFORMATION TECHNOLOGY AND CONTROL,
2006, Vol. 35, No.4 pp. 371-378.

And a fix for the distortion in the circles.

Turns out it was a stupid mistake with the calculation of k - I was using
(4/3) rather than (4.0/3.0), and hence getting integer division rather than
floating point division . . .

This now runs very nicely, well enough to actually successfully cut out my
filter plans (modulo a certain amount of drift from the paper shifting under
the blade).

A further update to the gcode handling.

This invludes a big clean up of the code, a much more complete parser for the
full gcode "spec", and some cleanups for the related code.

This is an interim commit to record status - it's not complete.

Fixes to the gcode handling.

It now at least approximately works, though it doesn't yet /draw/ anything in
the sim version thanks to not converting millimeter input to inches . . .

Fix the metric/imperial conversion.

I really want to change the internals to use metric, but this is fine as-is.

Cleanup and finalise gcode support.

This implementation is at least somewhat clean and includes some reasonable
support for debugging output, a reasonably clean parser, and support for all
the stuff that I've seen in any of the test gcode files I've looked at.

Additional features that I'm going to implement are G3 and useful support for
the G90/91 commands (though I have yet to see any files that would exercise
the non-default case).