Initial commit of gcode OO parsing/processing.
This changes the processing code to use an OO model that feels considerably
cleaner and better designed.

The design now keeps the parsing code in the gcode class, and moves the actual
drawing objects out to new classes, one per object type (line, bezier curve,
arc). Each type then has a draw() method that takes the cutter to be used, and
actually draws the relevant bits of the file.