Possibly finally working version of the g2 and g3 code.
This has been pretty painful to develop. I've been struggling to figure out
how to properly handle calculating the arc width, which should be a simple
matter of using atan2() but is complicated by the fact that sometimes that
method returns values that aren't valid (i.e. a positive value [anticlockwise] for a
calculation that should return a negative value [clockwise], or vice-versa).
How to determine which cases are what, and what to do with them, has been
difficult.

The current approach (where it's not valid, add/subtract 2PI to put it in the
right range) mostly works, but there are some scenarios where this is
problematic. These cases are where a circle has been subdivided into two arcs
that are complementary, but not equal - one is slightly more than pi, the other
slightly less. The current code results in the two arcs being the same, but
not complementary - either both are slightly more than pi, or both are
slightly less. Because g-code refers to arcs only, not circles, I have no way
of knowing that they /should/ be complementary aside from the actual values,
and I can't see how to figure it out from the pair of arcs.